+++ src/root/Root.tsx
@@ -31,6 +31,7 @@ import App from '../content/app/App';
 import PrivacyBanner from '../shared/components/privacy-banner/PrivacyBanner';
 import ErrorBoundary from 'src/shared/components/error-boundary/ErrorBoundary';
 import { GeneralErrorScreen } from 'src/shared/components/error-screen';
+import { setScrollbarWidth } from 'src/shared/helpers/scrollbarWidth';
 
 import styles from './Root.scss';
 
@@ -73,26 +74,6 @@ export const Root = () => {
       </ErrorBoundary>
     </div>
   );
-};
-
-// creating dummy elements to simulate scrollbar
-// Find scrollbar width and set it to css variable so that it can be used to calculate the right padding
-// Removing dummy elements once css variable is set
-const setScrollbarWidth = () => {
-  // Creating invisible container and child to show scrollbar
-  const outer = document.createElement('div');
-  const inner = document.createElement('div');
-
-  outer.style.overflow = 'scroll';
-
-  outer.appendChild(inner);
-  document.body.appendChild(outer);
-  const scrollbarWidth = outer.offsetWidth - outer.scrollWidth;
-
-  const rootNode = document.querySelector(':root') as HTMLElement;
-  rootNode?.style.setProperty('--scrollbar-width', `${scrollbarWidth}px`);
-
-  outer.parentNode?.removeChild(outer);
 };
 
 export default Root;
